docs: add SLAR documentation with paper citation and fix type consistency
- Add comprehensive docstring to Slar class with paper citation
  (Land & Gaalman, 1998) and algorithm explanation
- Document the `has_one` extension (not in original paper but kept
  for more aggressive starvation avoidance)
- Add inline comments explaining the two release triggers
- Fix type consistency: change `allowance_factor: int` to `float`
  across slar.py, lumscor.py, job.py, and builders.py
- Add null safety checks for `previous_server` in both SLAR and
  LumsCor starvation triggers (defensive programming)
- Fix docstring in builders.py: "Server Load Adjustment Rule" ->
  "Superfluous Load Avoidance Release"

Reference: https://doi.org/10.1016/S0925-5273(98)00052-8

+++ src/simulatte/job.py
@@ -287,7 +287,7 @@ class BaseJob(ABC):
             return self.due_date - window_size <= self.finished_at <= self.due_date + window_size
         raise ValueError("Job is not done or missing finish time. Cannot determine if finished in due date window.")
 
-    def planned_release_date(self, allowance: int = 2) -> SimTime:
+    def planned_release_date(self, allowance: float = 2.0) -> SimTime:
         """Calculate the optimal release time from the Pre-Shop Pool.
 
         Args:
